src/main/java/com/MusicStore/MusicStore/Controller/ReviewController.java
package com.MusicStore.MusicStore.Controller;

import com.MusicStore.MusicStore.Model.Review;
import com.MusicStore.MusicStore.Service.ReviewService;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("api/review")
public class ReviewController {

    private ReviewService reviewService;

    public ReviewController(ReviewService reviewService){
        this.reviewService = reviewService;
    }

    @PostMapping
    public ResponseEntity<Object> addReview(@RequestBody Review review){
        reviewService.addReview(review);
        return ResponseEntity.status(HttpStatus.CREATED).build();
    }

    @PutMapping
    public ResponseEntity<Object> updateReview(@RequestBody Review review){
        reviewService.updateReview(review);
        return ResponseEntity.ok().build();
    }

    @DeleteMapping
    public ResponseEntity<Object> deleteReview(@PathVariable Review review){
        reviewService.deleteReview(review);
        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
    }

    @GetMapping
    public ResponseEntity<List<Review>> getAllReviews(){
        return ResponseEntity.ok(reviewService.getAllReviews());
    }

}
